# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для записи к специалистам",
        "title": "LAPS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/specializations": {
            "get": {
                "description": "Возвращает список специализаций с фильтрацией и пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специализации"
                ],
                "summary": "Получить список специализаций",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Лимит записей на странице (по умолчанию 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Тип специалиста (lawyer, psychologist, nutritionist, trainer)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Фильтр по активности",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поисковый запрос",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID специалиста для фильтрации специализаций",
                        "name": "specialist_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список специализаций с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/rest.paginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новую специализацию (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специализации"
                ],
                "summary": "Создать специализацию",
                "parameters": [
                    {
                        "description": "Данные специализации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateSpecializationDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной специализации",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/specializations/{id}": {
            "get": {
                "description": "Возвращает информацию о специализации по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специализации"
                ],
                "summary": "Получить специализацию по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специализации",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные специализации",
                        "schema": {
                            "$ref": "#/definitions/domain.Specialization"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специализация не найдена",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о специализации (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специализации"
                ],
                "summary": "Обновить специализацию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специализации",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные специализации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateSpecializationDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном обновлении",
                        "schema": {
                            "$ref": "#/definitions/rest.messageResponseType"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специализация не найдена",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет специализацию (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специализации"
                ],
                "summary": "Удалить специализацию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специализации",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Специализация успешно удалена"
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специализация не найдена",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateSpecializationDTO": {
            "type": "object",
            "required": [
                "description",
                "name",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "lawyer",
                        "psychologist",
                        "nutritionist",
                        "trainer"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.SpecialistType"
                        }
                    ]
                }
            }
        },
...
        "domain.SpecialistType": {
            "type": "string",
            "enum": [
                "lawyer",
                "psychologist",
                "nutritionist",
                "trainer"
            ],
            "x-enum-varnames": [
                "SpecialistTypeLawyer",
                "SpecialistTypePsychologist",
                "SpecialistTypeNutritionist",
                "SpecialistTypeTrainer"
            ]
        },
        "domain.Specialization": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.SpecialistType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateSpecializationDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "rest.errorResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.messageResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.paginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}